Reorganizacion de codigo en archivo controlador
